+++ src/components/App/App.test.js
@@ -1,10 +1,19 @@
 import React from 'react';
-import App from './App';
-import {shallow, configure} from 'enzyme';
+import AppContainer from './AppContainer';
+import {shallow, configure, render} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
+import store from '../../utils/reduxStore';
+import {Provider} from 'react-redux';
+// import registerServiceWorker from './registerServiceWorker';
+
+// registerServiceWorker();
 
 configure({adapter: new Adapter()});
 
 it('renders without crashing', () => {
-	shallow(<App loadUsers={jest.fn} />);
+	shallow(
+		<Provider store={store}>
+			<AppContainer />
+		</Provider>
+	);
 });
