+++ src/components/StoreList/StoreListContainer.js
@@ -13,20 +13,6 @@ const mapStateToProps = state => {
 
 const mapDispatchToProps = dispatch => {
 	return {
-		getStoreList: () =>
-			getStore()
-				.then(items =>
-					items.map(item => ({
-						name:
-							item.name +
-							(item.qualifier ? ' ' + item.qualifier : ''),
-						index: item.id
-					}))
-				)
-				.then(storeList => {
-					dispatch(setStoreList(storeList));
-				})
-				.catch(() => dispatch(setLoadStoreListError(true))),
 		setShowList: showList => dispatch(setShowList(showList))
 	};
 };
