+++ src/index.js
@@ -3,39 +3,8 @@ import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import './index.css';
 import AppContainer from './components/App/AppContainer';
+import store from './utils/reduxStore';
 import registerServiceWorker from './registerServiceWorker';
-import {createStore, applyMiddleware, combineReducers, compose} from 'redux';
-import thunk from 'redux-thunk';
-import {
-	storeList,
-	loadStoreListError,
-	showList
-} from './components/StoreList/reducer';
-import {
-	users,
-	slackUserFetchError,
-	currentUser,
-	prediction
-} from './components/App/reducer';
-
-const rootReducer = combineReducers({
-	storeList,
-	loadStoreListError,
-	users,
-	slackUserFetchError,
-	currentUser,
-	showList,
-	prediction
-});
-
-const store = createStore(
-	rootReducer,
-	compose(
-		applyMiddleware(thunk),
-		window.__REDUX_DEVTOOLS_EXTENSION__ &&
-			window.__REDUX_DEVTOOLS_EXTENSION__()
-	)
-);
 
 ReactDOM.render(
 	<Provider store={store}>
